+++ src/vm/mem.c
@@ -75,7 +75,7 @@ mem_getWord(PmMemSpace_t memspace, uint8_t **paddr)
     uint8_t blo = mem_getByte(memspace, paddr);
     uint8_t bhi = mem_getByte(memspace, paddr);
 
-    return (uint16_t)(blo | (bhi << 8));
+    return (uint16_t)(blo | (bhi << (int8_t)8));
 }
 
 
@@ -86,7 +86,7 @@ mem_getInt(PmMemSpace_t memspace, uint8_t **paddr)
     uint16_t wlo = mem_getWord(memspace, paddr);
     uint32_t whi = mem_getWord(memspace, paddr);
 
-    return (uint32_t)(wlo | (whi << 16));
+    return (uint32_t)(wlo | (whi << (int8_t)16));
 }
 
 
@@ -125,7 +125,7 @@ mem_getNumUtf8Bytes(PmMemSpace_t memspace, uint8_t **psrc)
 {
     uint8_t *pbase = *psrc;
 
-    while (mem_getByte(memspace, psrc) != 0);
+    while (mem_getByte(memspace, psrc) != (uint8_t)0);
     return *psrc - pbase - 1;
 }
